[learning] merge handlers module
