+++ services/api/app/assistant/repositories/logs.py
@@ -210,27 +210,27 @@ def start_flush_task(interval: float = _FLUSH_INTERVAL) -> None:
         return
     _flush_task = loop.create_task(_flush_periodically(interval))
 
+    def _on_done(task: asyncio.Task[None]) -> None:
+        global _flush_task
+        exc: BaseException | None = None
+        try:
+            exc = task.exception()
+        except asyncio.CancelledError:  # pragma: no cover - expected
+            pass
+        if exc is not None:
+            logger.exception("Background flush task failed", exc_info=exc)
+            _flush_task = None
+            start_flush_task(interval)
+        else:
+            _flush_task = None
+
+    _flush_task.add_done_callback(_on_done)
+
 
 async def start_flush_task_async(interval: float = _FLUSH_INTERVAL) -> None:
     """Explicit API to start flush task from an async context."""
 
     start_flush_task(interval)
-
-        def _on_done(task: asyncio.Task[None]) -> None:
-            global _flush_task
-            exc: BaseException | None = None
-            try:
-                exc = task.exception()
-            except asyncio.CancelledError:  # pragma: no cover - expected
-                pass
-            if exc is not None:
-                logger.exception("Background flush task failed", exc_info=exc)
-                _flush_task = None
-                start_flush_task(interval)
-            else:
-                _flush_task = None
-
-        _flush_task.add_done_callback(_on_done)
 
 
 async def stop_flush_task() -> None:
