+++ services/api/app/diabetes/handlers/learn_handlers.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .learning_handlers import learn_command
+from ..learning_handlers import learn_command
 
 __all__ = ["learn_command"]
