+++ tests/diabetes/test_learning_logs.py
@@ -9,7 +9,7 @@ from telegram import Update
 from telegram.ext import CallbackContext
 
 from services.api.app.config import settings
-from services.api.app.diabetes.handlers import learning_handlers
+from services.api.app.diabetes import learning_handlers
 
 
 class DummyMessage:
@@ -35,6 +35,7 @@ def make_context(**kwargs: Any) -> CallbackContext[Any, Any, Any, Any]:
 async def test_lesson_start_logging(monkeypatch: pytest.MonkeyPatch, caplog: pytest.LogCaptureFixture) -> None:
     """Ensure starting a lesson logs start and completion events."""
     monkeypatch.setattr(settings, "learning_mode_enabled", True)
+    monkeypatch.setattr(settings, "learning_content_mode", "static")
     async def fake_ensure_overrides(update: object, context: object) -> bool:
         return True
 
@@ -66,6 +67,7 @@ async def test_lesson_start_logging(monkeypatch: pytest.MonkeyPatch, caplog: pyt
 async def test_exit_logging(monkeypatch: pytest.MonkeyPatch, caplog: pytest.LogCaptureFixture) -> None:
     """Ensure exiting a lesson logs start and completion events."""
     monkeypatch.setattr(settings, "learning_mode_enabled", True)
+    monkeypatch.setattr(settings, "learning_content_mode", "static")
     async def fake_run_db(*args: Any, **kwargs: Any) -> None:  # pragma: no cover - simple stub
         return None
