+++ tests/learning/test_handlers.py
@@ -170,7 +170,7 @@ async def test_static_mode_delegates(monkeypatch: pytest.MonkeyPatch) -> None:
         SimpleNamespace(learning_content_mode="static", learning_mode_enabled=True),
     )
     monkeypatch.setattr(
-        learning_handlers.legacy_handlers, "learn_command", fake_learn_command
+        learning_handlers, "_static_learn_command", fake_learn_command
     )
 
     upd = cast(Update, SimpleNamespace(message=object()))
